clean up alot
