+++ app/databases/redis/redis_client.py
@@ -1,7 +1,7 @@
 import redis
 from typing  import Union, Dict
 import json
-from constants import REDIS_SOLVED_KEY, REDIS_ATTEMPTED_KEY, TTL
+from constants import TTL
 
 # from util import get_secrets
 TTL = 300
@@ -77,20 +77,3 @@ class RedisClient:
         return self.client.ttl(key)
 
 
-    def check_if_user_has_attempted_problem(self, user_id: str, problem_difficulty: str):
-        try:
-            problem_difficulty_map = self.get_decoded_dict(REDIS_SOLVED_KEY)
-        except:
-            raise Exception("Invalid key: submission_key has to be either REDIS_SOLVED_KEY constant")
-        if problem_difficulty_map and problem_difficulty in problem_difficulty_map and user_id in problem_difficulty_map[problem_difficulty]:
-            return True
-        return False
-    
-    def check_if_user_has_submitted_problem(self, user_id: str, problem_difficulty: str):
-        try:
-            problem_difficulty_map = self.get_decoded_dict(REDIS_SOLVED_KEY)
-        except:
-            raise Exception("Invalid key: attempt_key has to be REDIS_ATTEMPTED_KEY constant")
-        if problem_difficulty_map and problem_difficulty in problem_difficulty_map and user_id in problem_difficulty_map[problem_difficulty]:
-            return True
-        return False
